refactor(ui): Extract utility functions and improve component modularity
- Move button variants to separate file `button-variant.ts`
- Create `navigation-menu-utils.ts` for navigation menu trigger styles
- Extract form field context and utility functions to `form-utils.ts`
- Add `icon-utils.ts` for icon component retrieval
- Simplify imports and component implementations
