+++ src/components/ui/form.tsx
@@ -7,24 +7,13 @@ import {
     FieldPath,
     FieldValues,
     FormProvider,
-    useFormContext,
 } from 'react-hook-form';
 
 import { cn } from '../../lib/utils';
 import { Label } from './label';
+import { useFormField, FormFieldContext, FormItemContext } from './form-utils';
 
 const Form = FormProvider;
-
-type FormFieldContextValue<
-    TFieldValues extends FieldValues = FieldValues,
-    TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-    name: TName;
-};
-
-const FormFieldContext = React.createContext<FormFieldContextValue>(
-    {} as FormFieldContextValue,
-);
 
 const FormField = <
     TFieldValues extends FieldValues = FieldValues,
@@ -38,37 +27,6 @@ const FormField = <
         </FormFieldContext.Provider>
     );
 };
-
-const useFormField = () => {
-    const fieldContext = React.useContext(FormFieldContext);
-    const itemContext = React.useContext(FormItemContext);
-    const { getFieldState, formState } = useFormContext();
-
-    const fieldState = getFieldState(fieldContext.name, formState);
-
-    if (!fieldContext) {
-        throw new Error('useFormField should be used within <FormField>');
-    }
-
-    const { id } = itemContext;
-
-    return {
-        id,
-        name: fieldContext.name,
-        formItemId: `${id}-form-item`,
-        formDescriptionId: `${id}-form-item-description`,
-        formMessageId: `${id}-form-item-message`,
-        ...fieldState,
-    };
-};
-
-type FormItemContextValue = {
-    id: string;
-};
-
-const FormItemContext = React.createContext<FormItemContextValue>(
-    {} as FormItemContextValue,
-);
 
 const FormItem = React.forwardRef<
     HTMLDivElement,
@@ -166,7 +124,6 @@ const FormMessage = React.forwardRef<
 FormMessage.displayName = 'FormMessage';
 
 export {
-    useFormField,
     Form,
     FormItem,
     FormLabel,
